fix(build): Cleanup dependency management (#273)
* fix(build): Cleanup dependency management

All dependencies were declared with the api scope which is not necessary and which leaks dependencies to downstream projects. After this change it should no longer be necessary for Grails apps to force the version of groovy-xml. Several unused dependencies have also been removed.

* chore(build): Add direct dependencies that are used in project

Project depends on grails-bootstrap and grails-gradle-model.
These were previously transitive dependencies pulled in by grails-shell
but as classes from them are used in the project they should be declared.